Migrate from Consent Management API Consent model to new GSI item model
Previously determined that the prior database integration would not
work as expected, and migrated the consent DynamoDB schema
ActiveConsentsByExpiryHour GSI that would support our use cases
of needing to efficiently look up all active consents past their
expiry date.

The ActiveConsentsByExpiryHour GSI will have a time-bucketed
expiryHour attribute as its GSI partition key, to enable
efficiently querying for all active consents with expiry times
that are in the given hour, and with an expiryTimeId attribute
as the GSI sort key, which appends expiryTime + "|" + the
consent partition key.

This sort key ensures that all GSI query results are automatically
sorted in ascending order of expiryTime, allowing us to efficiently
look up all consents with expiry times in the past for a given
expiryHour time bucket, while appending id to ensure that every
GSI partition key + sort key value is unique and identifies
a given consent record that we can update.

Every GSI item automatically includes the source DynamoDB's
keys, so we have the consent partition key id, and we also
specifically include the consentVersion attribute in the GSI
to allow us to increment the version in our update requests,
with a conditional write condition based on consent version
to mitigate concurrent writes that may otherwise result in
unexpected/corrupted data.

In this commit, migrating code to use the new DynamoDB GSI
data model, and migrating off the unnecessary dependency on the
Consent Management API Consent data model, which has more
attributes than are needed/available on the DynamoDB GSI that
we'll be querying.