Fix DynamoDB UpdateItemRequest for expiring consent
expireConsent requests were previously failing with error:
Invalid ConditionExpression: An expression attribute value used in expression is not defined; attribute value: :expectedLastVersion

This was due to the condition expression in the UpdateItemRequest
being a simple string, with any attribute names and values needing
to be included in expressionAttributeNames and
expressionAttributeValues.

Had specified these within the condition's Expression, however,
those are no longer used/relevant when parsing the Expression
into a string and passing into UpdateItemRequest builder's
conditionExpression(conditionExpression.expression()).